Wait for initial sync within  addProject()
